+++ src/lib/dom.js
@@ -35,58 +35,6 @@ lib.dom = {
 				lib.event.one("DOMContentLoaded", document, resolve);
 			}
 		});
-	},
-
-	getTransitionTime: new function () {
-
-		var style = document.documentElement.style,
-			transitionDuration = [
-				"transitionDuration",
-				"OTransitionDuration",
-				"MozTransitionDuration",
-				"WebkitTransitionDuration"
-			].find(isSupported),
-			transitionDelay = [
-				"transitionDelay",
-				"OTransitionDelay",
-				"MozTransitionDelay",
-				"WebkitTransitionDelay"
-			].find(isSupported);
-
-		style = null;
-
-		function isSupported(property) {
-			return property in style;
-		}
-
-		function parseTime(string) {
-			return string.split(",").map(function (time) {
-				return Number.parseInt(time, 10);
-			});
-		}
-
-		function getMaxTime(time1, time2) {
-			var i = 0, length = Math.max(time1.length, time2.length), time, maxTime = 0;
-			while (i < length) {
-				time = (time1[i] || 0) + (time2[i] || 0);
-				if (time > maxTime) {
-					maxTime = time;
-				}
-				i++;
-			}
-			return maxTime;
-		}
-
-		if (transitionDuration && transitionDelay) {
-			return function (element) {
-				var style = getComputedStyle(element);
-				return getMaxTime(parseTime(style[transitionDuration]), parseTime(style[transitionDelay])) * 1000;
-			};
-		}
-		return function () {
-			return 0;
-		};
-
 	}
 
 };
@@ -96,10 +44,16 @@ new function () {
 	function promise(element, method, classes) {
 		return new Promise(function (resolve) {
 			requestAnimationFrame(function () {
-				var delay, className = element.className;
+				var delay, style, className = element.className;
 				changeClassList(element.classList, method, classes);
 				if (className != element.className) {
-					delay = lib.dom.getTransitionTime(element);
+					style = getComputedStyle(element);
+					/*delay = Math.max(
+						lib.css.getTransitionTime(style),
+						lib.css.getFiniteAnimationTime(style)
+					);*/
+					delay = lib.css.getTransitionTime(style);
+					style = null;
 					if (delay) {
 						setTimeout(function () {
 							resolve(element);